В40 стратегии/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.10.11
websockets==12.0

# Trading
ccxt==4.4.89
ta==0.10.2

# Database
PyMySQL==1.1.0
sqlalchemy==2.0.23
redis==5.0.1
aiomysql>=0.1.1

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Notifications
python-telegram-bot==20.7

# Utils
python-dotenv==1.0.0
psutil==5.9.6
structlog==23.2.0
APScheduler==3.10.4

# ML (optional)
scikit-learn==1.3.2
tensorflow>=2.13.0
xgboost>=1.7.0
optuna>=3.3.0       # Hyperparameter optimization

# Дополнительные зависимости
httpx==0.25.2
asyncpg==0.29.0

scikit-learn==1.3.2
joblib==1.3.2




# Data Processing
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.28      # Technical indicators

# NLP
transformers>=4.30.0  # BERT/GPT models
nltk>=3.8.0
textblob>=0.17.0

# Social Media
tweepy>=4.14.0      # Twitter API
praw>=7.7.0         # Reddit API

# Visualization
plotly>=5.15.0
dash>=2.11.0

# News
feedparser>=6.0.10   # RSS feeds
newspaper3k>=0.2.8   # News parsing

# WebSocket
python-socketio>=5.9.0

scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.0
optuna>=3.3.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.28
transformers>=4.30.0
nltk>=3.8.0
textblob>=0.17.0
tweepy>=4.14.0
praw>=7.7.0
plotly>=5.15.0
dash>=2.11.0
feedparser>=6.0.10
newspaper3k>=0.2.8
python-socketio>=5.9.0
